feat: move tools to their own tab
