+++ package.json
@@ -25,7 +25,8 @@
     "react-dock": "^0.7.0",
     "react-router-dom": "^5.3.4",
     "use-debounce": "^10.0.2",
-    "video.js": "^7.21.3"
+    "video.js": "^7.21.3",
+    "react-intl": "^6.2.8"
   },
   "devDependencies": {
     "yaml": "^2.7.0",
@@ -90,7 +91,8 @@
     "react": "^18.2.0",
     "react-bootstrap": "1.6.6",
     "react-router-dom": "^5.3.4",
-    "video.js": "^7.21.3"
+    "video.js": "^7.21.3",
+    "react-intl": "^6.2.8"
   },
   "resolutions": {
     "strip-ansi": "6.0.1"
