+++ src/components/Slider.tsx
@@ -8,8 +8,8 @@ import {
 } from 'react';
 import videojs from 'video.js';
 import { useDebouncedCallback } from 'use-debounce';
-import { StashToolsConfig, useStashToolsConfig } from './useLocalForage';
-import { hooks, React } from './api';
+import { StashToolsConfig, useStashToolsConfig } from '../hooks';
+import { hooks, React } from '../api';
 import InteractiveAPI = PluginApi.hooks.InteractiveAPI;
 
 type PlayerElement = { player?: videojs.Player } | undefined;
