+++ src/components/index.tsx
@@ -1,0 +1,5 @@
+export * from './Slider';
+export * from './StrokeSlider';
+export * from './SyncSlider';
+export * from './InteractiveToolsTab';
+export * from './ScriptChooser';
