+++ src/pluginApi.d.ts
@@ -3,6 +3,11 @@
 declare namespace PluginApi {
   const React: typeof import('react');
   const ReactDOM: typeof import('react-dom');
+  type PropsWithChildren = React.PropsWithChildren;
+  type ReactElement = React.ReactElement;
+  type ReactNode = React.ReactNode;
+  type FC = React.FC;
+
   namespace GQL {
     const AddGalleryImagesDocument: { [key: string]: any };
     const AddTempDlnaipDocument: { [key: string]: any };
@@ -1579,15 +1584,28 @@ declare namespace PluginApi {
     };
   }
   export type ComponentNames = keyof typeof components | string;
+  export type PatchFunction<T extends any = any> = (
+    props: PropsWithChildren<T>,
+    ...args: any[]
+  ) => ReactNode;
   namespace patch {
-    function before(target: ComponentNames, fn: Function): void;
+    function before<T extends any = any>(
+      target: ComponentNames,
+      fn: PatchFunction<T>,
+    ): void;
 
-    function instead(target: ComponentNames, fn: Function): void;
+    function instead<T extends any = any>(
+      target: ComponentNames,
+      fn: PatchFunction<T>,
+    ): void;
 
-    function after(target: ComponentNames, fn: Function): void;
+    function after<T extends any = any>(
+      target: ComponentNames,
+      fn: PatchFunction<T>,
+    ): void;
   }
   namespace register {
-    function route(path: string, component: React.FC<any>): void;
+    function route(path: string, component: FC<any>): void;
   }
 }
